refactor: rename go to top button
